Extract agent scope projection read model
